requirements.txt
aniso8601==9.0.1
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
Flask==2.1.1
flask-restx==0.5.1
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.6.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.1
jsonschema==4.4.0
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
pyrsistent==0.18.1
pytest==7.1.1
python-decouple==3.6
pytz==2022.1
requests==2.27.1
six==1.16.0
tomli==2.0.1
urllib3==1.26.9
Werkzeug==2.1.1
zipp==3.7.0
